Rectify the coefficients for network features
The coefficients for SV+ and SV- were mistakenly switched in the software. Note this does not impact the results in the paper: we computed the feature values separately without the software (because the computation was too time-consuming for a single run by the software, especially for the blast feature that we evaluated in the paper), and apply the coefficients to calculate the final score.